Refactor ImportDate ActiveModel class
We want our active model classes to have a similar
behaviour to a active record. For this to work we need to
add the attributes on each child class that inherits from
our base active model class.

Also replace index action with show. It seems more appropiate
since we are passing a commodity code in the URL.