Refactor the ImportDate ActiveModel class
We needed to be able to prepopulate the date_select field
with the value from the session.

This is needed for when a user comes back to this question.
In addition we needed an invalid past date to be persisted
on the form so that users can see what they've done wrong.

Also updating the error message.