Define which step ids to remove from the session
Because some of the questions will be shared by
other routes, we can no longer rely on the position
of that question in the wizard.

The same question could be on position 4 or 8 depending
on which route the user lands on.

This change allows devs to define which questions to be
removed from the session while positioned on a particular step.
This way the STEP_ID now becomes purely a unique identifier
for that step.