Trade remedies becomes a pseudo steps
Because the trade remedies page now has logic in place
and we need to know which step to go back to depending
on the journey we are at, it needs to become a step.

This PR plumbs in trade remedies step on RoW to NI route,
when there is a trade defence in place.