Adds the CommodityContextService
We always pull out the commodity from a commodity_context to avoid
multiple queries for the commodity for the same arguments.

This is different from redis in the sense that the commodity_context is
only interesting for the current puma thread we're executing under and
reduces the amount of network requests the duty calculator has to make
by loading as much colocated data as possible.