Enable duty calculations on commodities with meursing measures
When a measure expresses meursing components and we have an answer from the user
about which code applies to their trade, we have enough information to
have the backend api surface the resolved measure components that we can
then use to execute duty calculations on the measure option that a
trader then take advantage of.

I've also updated the compound evaluator to express resolved meursing
duty expressions by concatenating the duty expression with a <br> tag
and the resolved duty expression coming back from the api response.