Deploy tag to staging when creating release
This occurs concurrently with the production deploy so whether to wait for it to complete is up to the developer.

In the normal 'click to release' process, this wont change anything - it will deploy a tag to staging which is the new release, the tag should be identical to the git sha already on Staging.

In a re-release scenario this creates the opportunity to check what we are about to re-release on Staging before we deploy it to production