HOTT-1481: Normalise user inputs with multipliers
When the user enters a unit who's input type is different from the type
used in the calculator we need to use the multiplier to retrieve the
normalised value that respects the units of the component (either
measure condition or measure component).

We've also migrated away from a dependence on massive fixture files with
difficult-to-configure attributes and are now instead using a commodity
factory.

Finally, we introduce a whole bunch of orchestrating factory traits that
make simulation of various conditions in the evaluators more straightforward

This is all required because we're simplifying the interface of the duty
calculator for the user such that they have a limited number of more
commonly known unit types to enter as part of the duty calculator
process.