HOTT-3621: Add healthcheckz to duty calculator
This adds a healthcheckz route that replaces the ping route. This will
be used when the app is deployed to ECS in line with the other apps.

I've also removed force_ssl as we're not using it, and modified the ci
config to use many of the jobs that were ported across to the orb.